Modified to use uint32_t & uint64_t instead of homegrown UINT/ULONG

+++ farsh.h
@@ -1,4 +1,5 @@
 #include <stddef.h>   /* for size_t */
+#include <stdint.h>   /* for uint32_t & uint64_t */
 
 /* Constants describing FARSH requirements to calling code */
 #define FARSH_MAX_HASHES             32  /* number of 32-bit hashes supported by the built-in key */
@@ -8,13 +9,13 @@
 
 
 /* Hash the buffer. The 'seed' value in any function below may be used to "personalize" the hash function. Use seed==0 if you don't need that. */
-unsigned farsh (const void *data, size_t bytes, unsigned long long seed);
+uint32_t farsh (const void *data, size_t bytes, uint64_t seed);
 
 /* Hash the buffer, computing 32*n bits long result, starting with hash number 'k' (k+n<=32), and write result to the buffer pointed by the hash parameter */
-void farsh_n (const void *data, size_t bytes, int k, int n, unsigned long long seed, void *hash);
+void farsh_n (const void *data, size_t bytes, int k, int n, uint64_t seed, void *hash);
 
 /* Hash the buffer with the user-supplied 16byte-aligned 1024-byte key material */
-unsigned farsh_keyed (const void *data, size_t bytes, const void *key, unsigned long long seed);
+uint32_t farsh_keyed (const void *data, size_t bytes, const void *key, uint64_t seed);
 
 /* Hash the buffer with the user-supplied 16byte-aligned, 1008+n*16 bytes long key material, and write 32*n bits long result to the buffer pointed by the hash parameter */
-void farsh_keyed_n (const void *data, size_t bytes, const void *key, int n, unsigned long long seed, void *hash);
+void farsh_keyed_n (const void *data, size_t bytes, const void *key, int n, uint64_t seed, void *hash);
